+++ find_duplicates/matching.py
@@ -187,7 +187,7 @@ def fuzzy_title_match(title, lang=None):
 #  - second (if not None) is based on swapping name order
 # --------------------------------------------------------------
 
-def get_author_tokens(author, decode_non_ascii=True):
+def get_author_tokens(author, decode_non_ascii=True, strip_initials=False):
     '''
     Take an author and return a list of tokens useful for duplicate
     hash comparisons. This function tries to return tokens in
@@ -196,8 +196,9 @@ def get_author_tokens(author, decode_non_ascii=True):
     '''
 
     if author:
-        # Leave ' in there for Irish names
-        remove_pat = re.compile(r'[,!@#$%^&*(){}`~"\s\[\]/]')
+        # Ensure Last,First is treated same as Last, First adding back space after comma.
+        comma_no_space_pat = re.compile(r',([^\s])')
+        author = comma_no_space_pat.sub(', \\1', author)
         replace_pat = re.compile(r'[-+.:;]')
         au = replace_pat.sub(' ', author)
         if decode_non_ascii:
@@ -206,16 +207,20 @@ def get_author_tokens(author, decode_non_ascii=True):
         if ',' in au:
             # au probably in ln, fn form
             parts = parts[1:] + parts[:1]
+        # Leave ' in there for Irish names
+        remove_pat = re.compile(r'[,!@#$%^&*(){}`~"\s\[\]/]')
+        # We will ignore author initials of only one character.
+        min_length = 1 if strip_initials else 0
         for tok in parts:
             tok = remove_pat.sub('', tok).strip()
-            if len(tok) > 0 and tok.lower() not in IGNORE_AUTHOR_WORDS_MAP:
+            if len(tok) > min_length and tok.lower() not in IGNORE_AUTHOR_WORDS_MAP:
                 yield tok.lower()
 
 def identical_authors_match(author):
     return author.lower(), None
 
 def similar_authors_match(author):
-    author_tokens = list(get_author_tokens(author))
+    author_tokens = list(get_author_tokens(author, strip_initials=True))
     ahash = ' '.join(author_tokens)
     rev_ahash = None
     if len(author_tokens) > 1:
@@ -536,7 +541,10 @@ def do_assert_tests():
     assert_author_match('similar', 'authors', 'China Miéville', 'China Mieville')
     assert_author_match('similar', 'authors', 'Kevin Anderson', 'Anderson Kevin')
     assert_author_match('similar', 'authors', 'Kevin, Anderson', 'Anderson, Kevin')
-    assert_author_nomatch('similar', 'authors', 'Kevin J. Anderson', 'Kevin Anderson')
+    assert_author_match('similar', 'authors', 'Kevin J. Anderson', 'Anderson,Kevin J.')
+    assert_author_match('similar', 'authors', 'Kevin Anderson', 'Anderson,Kevin J.')
+    assert_author_match('similar', 'authors', 'Kevin Anderson', 'Anderson,Kevin J')
+    assert_author_nomatch('identical', 'authors', 'Kevin, Anderson', 'Anderson, Dr Kevin')
 
     # Test our soundex author algorithms
     assert_author_match('soundex', 'authors', 'Kevin J. Anderson', 'Kevin J. Anderson')
@@ -595,7 +603,7 @@ def do_assert_tests():
 
 
 # For testing, run from command line with this:
-# calibre-debug -e algorithms.py
+# calibre-debug -e matching.py
 if __name__ == '__main__':
     do_assert_tests()